布局计算数据类型替换

+++ harmony/flash_list/src/main/cpp/AutoLayoutShadow.h
@@ -1,3 +1,27 @@
+/**
+ * MIT License
+ *
+ * Copyright (C) 2024 Huawei Device Co., Ltd.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 //
 // Created on 5/3/2024.
 //
@@ -23,9 +47,9 @@ namespace rnoh {
 
     public:
         bool horizontal{false};
-        int scrollOffset{0};
-        int offsetFromStart{0};
-        int windowSize{0};
+        Float scrollOffset{0};
+        Float offsetFromStart{0};
+        Float windowSize{0};
         Float renderOffset{0};
         Float blankOffsetAtStart{0};  // Tracks blank area from the top
         Float blankOffsetAtEnd{0};    // Tracks blank area from the bottom
@@ -34,7 +58,7 @@ namespace rnoh {
         AutoLayoutShadow(){};
         ~AutoLayoutShadow(){};
         void clearGapsAndOverlaps(std::vector<CellContainerComponentInstance::Shared> sortedItems);
-        int computeBlankFromGivenOffset(int actualScrollOffset, int distanceFromWindowStart, int distanceFromWindowEnd);
+        Float computeBlankFromGivenOffset(Float actualScrollOffset, Float distanceFromWindowStart, Float distanceFromWindowEnd);
     };
 } // namespace rnoh
